Feature/krmongo

+++ kryptonic/krmongo/mongo_client.py
@@ -3,6 +3,7 @@ from collections import namedtuple
 from contextlib import contextmanager
 from pymongo import MongoClient as BaseMongoClient
 from pymongo import monitoring
+from kryptonic.krmongo.auditlogging import LogEntry
 
 _kr_client = None
 
@@ -27,15 +28,14 @@ class KryptonicEvents(monitoring.CommandListener):
         pass
 
     def _log_insert(self, event):
-        entries = [{
-            'krData': {
-                'runId': self.run_id,
-                'db': event.database_name,
-                'collection': event.command['insert'],
-                'lastModified': datetime.utcnow()
-            },
-            'document': document
-        } for document in event.command['documents']]
+
+        entries = [LogEntry(
+            runId=self.run_id,
+            action='insert',
+            database=event.database_name,
+            collection=event.command['insert'],
+            document=document).values
+            for document in event.command['documents']]
 
         self.kr_client.get_database(KRYPTONIC_DATABASE).get_collection(KRYPTONIC_WRITES_COLLECTION).insert_many(entries)
 
@@ -71,14 +71,16 @@ def krmongorun(run_id=None, *args, **kwargs):
     if run_id is None:
         run_id = _generate_run_id()
 
-    monitoring.register(KryptonicEvents(run_id, kr_client, *args, **kwargs))
-    kr_client = BaseMongoClient(*args, **kwargs)
+    # kr_client = BaseMongoClient(*args, **kwargs)
+    # monitoring.register(KryptonicEvents(run_id, , *args, **kwargs))
+    mongo_client = MongoClient(run_id=run_id, *args, **kwargs)
     try:
-        yield RunInfo(run_id)
+        yield mongo_client
+        # yield RunInfo(run_id)
     except KrMongoNoCleanup:
         pass
     else:
-        _cleanup(kr_client, run_id)
+        mongo_client.cleanup(run_id)
 
 
 def _generate_run_id():
@@ -90,13 +92,14 @@ def _cleanup(client: MongoClient, run_id, kr_database=KRYPTONIC_DATABASE, writes
     Removes all documents that were recorded in kryptonic.transactions with a particular run_id, then removes the transaction records
     """
 
-    to_remove = client[kr_database].get_collection(writes_collection).find({'krData.runId': run_id})
+    to_remove = client[kr_database].get_collection(writes_collection).find({'runId': run_id})
     kr_ids_to_delete = []
-
     for doc in to_remove:
-        database = doc['krData']['db']
-        collection = doc['krData']['collection']
+        database = doc['database']
+        collection = doc['collection']
         doc_id = doc['document']['_id']
         client.get_database(database).get_collection(collection).delete_one({'_id': doc_id})
         kr_ids_to_delete.append({'_id': doc['_id']})
-    client.get_database(kr_database).get_collection(writes_collection).delete_many({'$or': kr_ids_to_delete})
+
+    if kr_ids_to_delete:
+        client.get_database(kr_database).get_collection(writes_collection).delete_many({'$or': kr_ids_to_delete})
